ZEUS-2475: Move SettingsStore migrations to separate file
